app,taskを廃止しworkflowに一本化
理由:
taskはappに従属的すぎた。
ならばworkflowという枠組みの中でtaskはそのサブモジュールというかたちにしたほうが実態を表現することができる。

---

変更内容

  1. 新しいモジュール構造
    - src/workflow/news.rs (Rust 2018方式に準拠)
    - src/workflow/news/article.rs
    - src/workflow/news/rss.rs
    - src/workflow/news/fixtures/
  2. 関数の改名と改善
    - execute_rss_workflow → workflow_news
    - 引数をgroup: Option<&str>からquery: Option<RssLinkQuery>に変更
    - より柔軟なクエリ指定が可能
  3. モジュール再編
    - src/app/ディレクトリを削除
    - src/task/ディレクトリを削除
    - src/lib.rsでworkflowモジュールを公開
  4. 依存関係修正
    - RssLinkQueryにCloneトレイトを追加
    - 全テスト96件が成功
    - Clippyリントもクリア
