--- src/app/mod.rs
@@ -1,4 +1,0 @@
-pub mod workflow;
-
-// workflowモジュールから公開
-pub use workflow::execute_rss_workflow;
