+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod app;
 pub mod core;
 pub mod infra;
-pub mod task;
+pub mod workflow;
