--- src/task/mod.rs
@@ -1,5 +1,0 @@
-pub mod article;
-pub mod rss;
-
-pub use article::task_collect_articles;
-pub use rss::task_collect_article_links;
