+++ src/workflow/mod.rs
@@ -1,0 +1,3 @@
+pub mod news;
+
+pub use news::workflow_news;
