+++ src/workflow/news/fixtures/article.sql
