+++ src/workflow/news/fixtures/article_mixed.sql
