+++ src/workflow/news/fixtures/concurrent_processing.sql
